Optimization for Material theme / tabbed: don't render navs of other tabs
When a tab has a section index page, and it is not the currently selected tab, skip rendering its sub-items entirely.

On desktop view this makes no difference, but in mobile view you will no longer be able to navigate out of the current tab and directly into sub-items of another tab -- first you'll have to navigate to the index page of the other tab.

So this is basically a downgrade, but for large sites this is also a huge optimization in terms of build time and disk usage.